pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "gopro-overlay"
version = "0.133.0"
description = "Overlay graphics dashboards onto GoPro footage, or create videos from GPX or FIT files"
readme = "README-PYPI.md"
requires-python = ">=3.11"
license = "GPL-3.0-or-later"
authors = [
    {name = "James Richardson", email = "james+gopro@time4tea.net"}
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.11",
    "Environment :: Console",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "pillow==11.3.0",
    "pint==0.25.1",      # min python version 3.11
    "gpxpy==1.6.1",
    "fitdecode==0.10.0",
    "geographiclib==1.52",
    "progressbar2==4.2.0",
    "requests==2.32.5",
    "sqlitedict==2.1.0",
    "aiohttp==3.13.3",   # for geotiler
    "cytoolz==1.1.0",    # for geotiler
]

[project.urls]
Homepage = "https://github.com/time4tea/gopro-dashboard-overlay"
Source = "https://github.com/time4tea/gopro-dashboard-overlay"

[project.optional-dependencies]
cairo=["pycairo==1.28.0"]
test = ["gopro-overlay[cairo]", "pytest", "pytest-asyncio"]
dev = ["gopro-overlay[test]", "black", "viztracer", "bump-my-version"]

[tool.setuptools.packages.find]
where = ["."]
include = ["gopro_overlay*"]

[tool.setuptools.package-data]
"gopro_overlay" = ["**/*.xml", "**/*.json"]

[tool.setuptools]
include-package-data = true

script-files= [
    "bin/gopro-contrib-data-extract.py",
    "bin/gopro-cut.py",
    "bin/gopro-dashboard.py",
    "bin/gopro-extract.py",
    "bin/gopro-join.py",
    "bin/gopro-layout.py",
    "bin/gopro-rename.py",
    "bin/gopro-to-csv.py",
    "bin/gopro-to-gpx.py",
    "bin/gopro-debug.py",
]

